# koicare/koi_swagger.yaml
openapi: 3.0.1
info:
  title: Koi Care API
  description: API documentation for your web application
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /koi/{id}:
    get:
      summary: Get koi by ID
      tags:
        - Koi
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      responses:
        "200":
          description: Koi retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  image:
                    type: string
                  body_shape:
                    type: string
                  age:
                    type: integer
                  size:
                    type: number
                    format: float
                  weight:
                    type: number
                    format: float
                  gender:
                    type: string
                  breed:
                    type: string
                  origin:
                    type: string
                  selling_price:
                    type: number
                    format: float
                  pond_id:
                    type: integer
        "404":
          description: Koi not found

    put:
      summary: Update koi by ID
      tags:
        - Koi
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                body_shape:
                  type: string
                age:
                  type: integer
                size:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
                gender:
                  type: string
                breed:
                  type: string
                origin:
                  type: string
                selling_price:
                  type: number
                  format: float
                pond_id:
                  type: integer
      responses:
        "200":
          description: Koi updated successfully
        "404":
          description: Koi not found

    delete:
      summary: Delete koi by ID
      tags:
        - Koi
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi
      responses:
        "200":
          description: Koi deleted successfully
        "404":
          description: Koi not found
  /koiGrowth/{id}:
    get:
      summary: Get koi growth record by ID
      tags:
        - Koi Growth
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi growth record
      responses:
        "200":
          description: Koi growth record retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  growth_date:
                    type: string
                    format: date
                  age:
                    type: integer
                  size:
                    type: number
                    format: float
                  weight:
                    type: number
                    format: float
                  koi_id:
                    type: integer
        "404":
          description: Koi growth record not found

    put:
      summary: Update koi growth record by ID
      tags:
        - Koi Growth
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the koi growth record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                growth_date:
                  type: string
                  format: date
                age:
                  type: integer
                size:
                  type: number
                  format: float
                weight:
                  type: number
                  format: float
      responses:
        "200":
          description: Koi growth record updated successfully
        "404":
          description: Koi growth record not found
...
  /order:
    post:
      summary: Create a new order
      tags:
        - Order
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:  
              type: object
              properties:
                orderItems:
                  type: array
                items:
                  type: object
                  properties: 
                    product_id:
                      type: integer
                    quantity:
                      type: integer
                    price:
                      type: number
                      format: float
                totalAmount:
                  type: number
                  format: float
      responses:
        '201':
          description: Order created successfully
        '500':
          description: Internal server error
  /order/{id}:
    get:
      summary: Get order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: integer
                  user_id:
                    type: integer
                  order_date:
                    type: string
                    format: date-time
                  total_amount:
                    type: number
                    format: float
                  status:
                    type: string
                    enum: [pending, processing, shipped, delivered, cancelled]
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id:
                          type: integer
                        name:
                          type: string
                        price:
                          type: number
                          format: float
                        quantity:
                          type: integer
        '404':
          description: Order not found

    put:
      summary: Update order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderItems:
                  type: array
                  items:
                    type: object
                    properties:  
                      product_id:
                        type: integer
                      quantity:
                        type: integer
                      price:
                        type: number
                        format: float
                totalAmount:
                  type: number
                  format: float
                status:
                  type: string
                  enum: [pending, processing, shipped, delivered, cancelled]
      responses:
        '200':
          description: Order updated successfully
        '404':
          description: Order not found

    delete:
      summary: Delete order by ID
      tags:
        - Order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the order
      responses:
        '200':
          description: Order deleted successfully
        '404':
          description: Order not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT